+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v1.5.1 [unreleased]
 
 - [#1982](https://github.com/influxdata/kapacitor/pull/1982): Fix KafkaTopic not working from TICKscript
+- [#1989](https://github.com/influxdata/kapacitor/pull/1989): Improve Kafka alert throughput.
 
 ## v1.5.0 [2018-05-17]